+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - RowBinary: truncate NaiveDateTime resulting from DateTime64 https://github.com/plausible/ch/pull/297
 - Add support for `JSON(...)` (JSON with options) type https://github.com/plausible/ch/pull/309
 - RowBinary: de- and encode dynamic JSON https://github.com/plausible/ch/pull/296
+- use gregorian seconds for naive datetime encoding in rowbinary (it's faster this way) https://github.com/plausible/ch/pull/311
 
 ## 0.7.1 (2026-01-15)